crash: Disable SNP on crashkernel kexec
The crashkernel is tasked with two problems as it stands right now:
During start it will read the SNP enabled IOMMU tables during early init
which can cause the lapic configuration to be configured into invalid or
unintuitive modes.
This will then cause the crashkernel to panic due to the lapic timer calibration
failing or general RCU stalling.
As the crashkernel can not pick up the host kernels PSP/SEV/SNP configuration
we are therefore disabeling SNP not only on the normal kexec path but also
on crash-kexec.

This needs to be called very early on in the machine teardown as it requires
to set per-cpu state.

Signed-off-by: Mimoja <git@mimoja.de>
Signed-off-by: Johanna 'Mimoja' Amelie Schander <mimoja@amazon.com>

Reviewed-by: Alexander Graf (AWS) <graf@amazon.de>
Reviewed-by: Costin Lupu <lvpv@amazon.nl>
Reviewed-by: Roman Kagan <rkagan@amazon.de>